Omit 2D annotations from dataset and result uploads
