+++ examples/dataset/object_detection_3d/object_detection_3d/upload_dataset.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 from argparse import ArgumentParser
 from argparse import Namespace
-from collections import Counter
 
 import pandas as pd
 from object_detection_3d.constants import BUCKET
@@ -21,13 +20,9 @@ from object_detection_3d.constants import DATASET
 from object_detection_3d.constants import DEFAULT_DATASET_NAME
 from object_detection_3d.constants import ID_FIELDS
 from object_detection_3d.constants import TASK
-from object_detection_3d.utils import create_velo_to_pixel_matrix
+from object_detection_3d.utils import load_data
 
-from kolena.annotation import LabeledBoundingBox
-from kolena.annotation import LabeledBoundingBox3D
-from kolena.asset import ImageAsset
 from kolena.dataset import upload_dataset
-from kolena.workflow.asset import PointCloudAsset
 
 # KITTI only supports evaluation of the first three classes but "Van" and "Person_sitting" GTs
 # are used to avoid penalizing inferences labeled as "Car" and "Pedestrian" respectively.
@@ -48,77 +43,6 @@ LABEL_FILE_COLUMNS = {
 }
 
 
-def load_data(df_raw: pd.DataFrame) -> pd.DataFrame:
-    records = []
-    meta_cols = [col for col in df_raw.columns if col not in LABEL_FILE_COLUMNS]
-
-    for record in df_raw.itertuples():
-        bboxes_2d = [
-            LabeledBoundingBox(
-                label=box["label"],
-                top_left=tuple(box["left_top_left"]),
-                bottom_right=tuple(box["left_bottom_right"]),
-                occluded=box["occluded"],  # type: ignore[call-arg]
-                truncated=box["truncated"],  # type: ignore[call-arg]
-                alpha=box["alpha"],  # type: ignore[call-arg]
-                difficulty=box["difficulty"],  # type: ignore[call-arg]
-            )
-            for box in record.objects
-        ]
-        bboxes_3d = [
-            LabeledBoundingBox3D(
-                label=box["label"],
-                dimensions=(box["dim_x"], box["dim_y"], box["dim_z"]),
-                center=(box["loc_x"], box["loc_y"], box["loc_z"] + (box["dim_z"] / 2.0)),
-                # translate in z (up) axis to be in center
-                rotations=(0.0, 0.0, box["rotation_y"]),  # Z yaw axis for lidar coordinates
-                occluded=box["occluded"],  # type: ignore[call-arg]
-                truncated=box["truncated"],  # type: ignore[call-arg]
-                alpha=box["alpha"],  # type: ignore[call-arg]
-                difficulty=box["difficulty"],  # type: ignore[call-arg]
-            )
-            for box in record.objects
-        ]
-        counts = Counter([r["label"] for r in record.objects])
-        metadata = {col: getattr(record, col) for col in meta_cols}
-
-        records.append(
-            {
-                "image_id": record.image_id,
-                "locator": record.left_image,
-                "image_bboxes": bboxes_2d,
-                "images": [
-                    ImageAsset(
-                        locator=record.left_image,
-                        side="left",  # type: ignore[call-arg]
-                        # type: ignore[call-arg]
-                        projection=create_velo_to_pixel_matrix(record.Tr_velo_to_cam, record.P2),
-                        velodyne_bboxes=bboxes_3d,  # type: ignore[call-arg]
-                    ),
-                    ImageAsset(
-                        locator=record.right_image,
-                        side="right",  # type: ignore[call-arg]
-                        # type: ignore[call-arg]
-                        projection=create_velo_to_pixel_matrix(record.Tr_velo_to_cam, record.P3),
-                        velodyne_bboxes=bboxes_3d,  # type: ignore[call-arg]
-                    ),
-                ],
-                "velodyne": PointCloudAsset(locator=record.velodyne),
-                "total_objects": len(bboxes_3d),
-                "n_car": counts["Car"],
-                "n_pedestrian": counts["Pedestrian"],
-                "n_cyclist": counts["Cyclist"],
-                "velodyne_bboxes": bboxes_3d,
-                "velodyne_to_camera_transformation": record.Tr_velo_to_cam,
-                "camera_rectification": record.R0_rect,
-                "projection": create_velo_to_pixel_matrix(record.Tr_velo_to_cam, record.P2),
-                **metadata,
-            },
-        )
-
-    return pd.DataFrame(records)
-
-
 def run(args: Namespace) -> int:
     df_raw = pd.read_json(
         f"s3://{BUCKET}/{DATASET}/{TASK}/raw/{DATASET}.jsonl",
@@ -128,6 +52,7 @@ def run(args: Namespace) -> int:
         storage_options={"anon": True},
     )
     df = load_data(df_raw)
+    df.drop(columns=["image_bboxes"], inplace=True)
     upload_dataset(args.dataset, df, id_fields=ID_FIELDS)
     return 0
